move configcheck in pkgconfig package
