+++ pkgconfig/configchecker_test.go
@@ -1,4 +1,4 @@
-package pkgutils
+package pkgconfig
 
 import (
 	"os"
