+++ pkgconfig/loggers.go
@@ -591,5 +591,5 @@ func (c *ConfigLoggers) SetDefault() {
 }
 
 func (c *ConfigLoggers) IsValid(userCfg map[string]interface{}) error {
-	return CheckConfig(reflect.ValueOf(*c), userCfg)
+	return CheckConfigWithTags(reflect.ValueOf(*c), userCfg)
 }
